fix warning

+++ libmini/mini/miniio.cpp
@@ -58,7 +58,7 @@ void writefile(const char *filename,unsigned char *data,long long bytes)
    if (bytes<1) ERRORMSG();
 
    if ((file=fopen(filename,"wb"))==NULL) IOERROR();
-   if (fwrite(data,1,bytes,file)!=bytes) IOERROR();
+   if ((long long)fwrite(data,1,bytes,file)!=bytes) IOERROR();
 
    fclose(file);
    }
